fixed ALLLLLL bugs

+++ src/main/java/javathon16/Models/Client.java
@@ -9,9 +9,9 @@ public class Client {
 
     @Id
     @GeneratedValue
-    private Long id;
+    private int id;
 
-    private int phoneNumber;
+    private Long phoneNumber;
 
     private String name;
 
@@ -20,26 +20,25 @@ public class Client {
 
     private Client(){}
 
-    public Client(int phoneNumber, String name){
+    public Client(Long phoneNumber, String name){
         this.phoneNumber = phoneNumber;
         this.name = name;
         this.bonuses = new HashSet<>();
-
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public int getPhoneNumber() {
+    public Long getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(Long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
