+++ src/main/java/javathon16/Repositories/BonusRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface BonusRepository extends CrudRepository<Bonus, Long> {
+public interface BonusRepository extends CrudRepository<Bonus, Integer> {
 
-    List<Bonus> findAllByClientIdAndCompanyId(Long clientId, Long companyId);
+    List<Bonus> findAllByClientIdAndCompanyId(int clientId, int companyId);
 
 }
